# docs/swagger.yaml
basePath: /
definitions:
  handler.createEventInput:
    properties:
      date:
        type: string
      description:
        type: string
      location:
        type: string
      max_participants:
        minimum: 0
        type: integer
      title:
        type: string
    required:
    - date
    - title
    type: object
  handler.signInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.signUpInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for the GoEvent student event management platform.
  title: GoEvent API
  version: "1.0"
paths:
  /api/v1/events:
    get:
      consumes:
      - application/json
      description: get list of events with optional filtering and pagination
      parameters:
      - default: 10
        description: limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: offset
        in: query
        name: offset
        type: integer
      - description: sort field
        enum:
        - id
        - title
        - date
        - location
        in: query
        name: sort_by
        type: string
      - description: order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: filter by title
        in: query
        name: title
        type: string
      - description: filter by location
        in: query
        name: location
        type: string
      - description: filter from date (RFC3339)
        in: query
        name: from_date
        type: string
      - description: filter to date (RFC3339)
        in: query
        name: to_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: List Events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: create a new event
      parameters:
      - description: event info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.createEventInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create Event
      tags:
      - events
  /api/v1/events/{id}:
    delete:
      consumes:
      - application/json
      description: delete an event (creator only)
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete Event
      tags:
      - events
    get:
      consumes:
      - application/json
      description: get a single event by id
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Event By ID
      tags:
      - events
    put:
      consumes:
      - application/json
      description: update an event (creator only)
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: event info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.createEventInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update Event
      tags:
      - events
  /api/v1/events/{id}/register:
    post:
      description: register user for an event
      operationId: register-event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Register for Event
      tags:
      - events
  /api/v1/events/{id}/unregister:
    delete:
      description: unregister user from an event
      operationId: cancel-event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Cancel Event Registration
      tags:
      - events
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
      summary: Sign In
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create account
      operationId: create-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signUpInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Sign Up
      tags:
      - auth
swagger: "2.0"